IC-90 Changed setting the category from dropdown to search.
To implement this the control to set the value is changed, to block to prettify the overview is changed and the matching with a product.

To clarify where the category is placed the path is added in the search result, for cases where there are categories with the same name in different paths.